* 'master' of https://github.com/vistalab/mrQ:
  fix a bug in acpc alignment
  added templates to mrQ
  minor changes

Conflicts:
	SPGR/SPGR_pre/mrQ_initSPGR.m
	batch/mrQ_run.m

+++ batch/mrQ_run.m
@@ -117,7 +117,7 @@ function mrQ_run(mrQfileName,clobber)
 if (~exist(mrQfileName,'file') )
     error(['cant find ' mrQfileName ])
 else
-load (mrQfileName) 
+    load (mrQfileName)
 
 end
 
@@ -153,9 +153,9 @@ sub = mrQ.sub;
 
 
 if isfield(mrQ,'Arange_Date') && ~isempty(mrQ.Arange_Date)
-    
+
     fprintf([ '\n  using data intiate at ' mrQ.Arange_Date ' !  \n']);
-    
+
 else
     fprintf('\n  Arranging the data ...\n  ')
     [mrQ]=  mrQ_arrangeData(mrQ.RawDir,mrQ.arrangeRawFlag,mrQ.SEIR_seriesNumbers,mrQ.SPGR_seriesNumbers,mrQ.channels,mrQ.useNiftiFlag,mrQ);
@@ -181,18 +181,18 @@ else
 end
 
 if (mrQ.SEIR_done==0);
-    
+
     %keep track of the variable we use  for detail see inside the function
     [~, ~, ~, mrQ.SEIRsaveData]=mrQ_initSEIR(mrQ,mrQ.SEIRepiDir,mrQ.alignFlag,mrQ.complexFlag,mrQ.useAbs);
-    
+
     [mrQ]=mrQ_fitSEIR_T1(mrQ.SEIRepiDir,[],[],0,mrQ);
     mrQ.SEIR_done=1;
     save(mrQ.name,'mrQ');
     fprintf('fit SEIR  - done!');
-    
+
 else
     fprintf('\n  load fit SEIR data ! \n');
-    
+
 end
 
 %% intiate and  Align SPGR
@@ -207,16 +207,16 @@ else
 end
 
 if     mrQ.SPGR_init_done==0
-    
+
     %keep track of the variable we use  for detail see inside the function
     [~, ~, ~,~,~, mrQ]=mrQ_initSPGR(mrQ.SPGR,mrQ.refIm,mrQ.mmPerVox,mrQ.interp,mrQ.skip,[],mrQ);
     mrQ.SPGR_init_done=1;
-    
+
     save(mrQ.name,'mrQ');
     fprintf('\n  init SPGR - done!           \n');
 else
     fprintf(' \n load init SPGR data            \n');
-    
+
 end
 
 
@@ -241,15 +241,15 @@ end
 
 
 if  (mrQ.coilWeights==1 && mrQ.coilNum(1)>8  && mrQ.SPGR_coilWeight_done==0)
-    
+
     fprintf('\n Determining optimal coil weighting...\n');
     % Should this return the new structure with the weighting applied?
     [mrQ.AligndSPGR]=mrQ_multicoilWeighting(mrQ.spgr_initDir,mrQ.SPGR_niiFile,mrQ.SPGR_niiFile_FA,mrQ);
     mrQ.SPGR_coilWeight_done=1;
     fprintf('\n SPGR  coil weighting - done!               \n');
-    
+
 else
-    
+
 end
 save(mrQ.name,'mrQ');
 
@@ -265,12 +265,12 @@ end
 if (mrQ.SPGR_T1fit_done==0);
     [mrQ.AnalysisInfo]=mrQfit_T1M0_ver2(mrQ);
     mrQ.SPGR_T1fit_done=1;
-    
+
     save(mrQ.name,'mrQ');
     fprintf('\n fit T1 SPGR  - done!              \n');
 else
     fprintf('\n load fited  SPGR T1                \n');
-    
+
 end
 save(mrQ.name,'mrQ');
 
@@ -286,7 +286,7 @@ else
 end
 
 if     mrQ.brakeAfterT1==1
-        fprintf('\n brake T1 map  are done              \n');
+    fprintf('\n brake T1 map  are done              \n');
 
     return
 end
@@ -300,24 +300,24 @@ else
 end
 
 if mrQ.segmentaion==0;
-   
-   
-       % run Free surfare
-     if (mrQ.runfreesurfer==1)
+
+
+    % run Free surfare
+    if (mrQ.runfreesurfer==1)
         mrQ=mrQ_Complitfreesurfer(mrQ);
-         
+
         mrQ.segmentaion=1;
         % use an uploaded freesurafre nii.zg
-     elseif isfield(mrQ,'freesurfer');
-         [mrQ.AnalysisInfo]=mrQ_CSF(mrQ.spgr_initDir,mrQ.freesurfer,[],mrQ.AnalysisInfo);
+    elseif isfield(mrQ,'freesurfer');
+        [mrQ.AnalysisInfo]=mrQ_CSF(mrQ.spgr_initDir,mrQ.freesurfer,[],mrQ.AnalysisInfo);
 
-              mrQ.segmentaion=1;
+        mrQ.segmentaion=1;
 
     else
         % Segment the T1w by FSL (step 1) and get the tissue mask (CSF WM GM) (step 2)
         mrQ=mrQ_segmentT1w2tissue(mrQ);
         mrQ.segmentaion=1;
-        
+
     end
     save(mrQ.name,'mrQ');
 end
@@ -330,7 +330,7 @@ end
 
 if (mrQ.calM0_done==0);
     fprintf('\n calculate M0 for each coil               \n');
-    
+
     %build a multi coil M0 image for the coils raw data and then fitted T1
     [mrQ.M0combineFile] = mrQ_multicoilM0(mrQ.spgr_initDir,[],[],mrQ.SPGR_niiFile,mrQ.SPGR_niiFile_FA,mrQ);
 
@@ -338,7 +338,7 @@ if (mrQ.calM0_done==0);
     save(mrQ.name,'mrQ');
 else
     fprintf('\n load M0 of each coil               \n');
-    
+
 end
 
 %
@@ -353,8 +353,8 @@ end
 
 if mrQ.SPGR_PDfit_done==0;
     fprintf('\n calculate PD from the M0 of all the coil               \n');
-    
-   
+
+
     %send the a call for the grid to fit  PD
 %  [mrQ.opt]=mrQ_fitPD_multicoil(mrQ.spgr_initDir,1,[],mrQ.PolyDeg,[],mrQ.sub,mrQ.proclass);
  if isfield(mrQ,'opt_logname');
@@ -365,18 +365,18 @@ mrQ_fitM0boxesCall(mrQ.opt_logname,mrQ.SunGrid,mrQ.proclus);
 
 
  end
- 
+
     if mrQ.SunGrid==1;
         %check for SGE is done  before you move on (while loop)
         mrQ.SPGR_PDfit_done=mrQ_Gridcheack(mrQ.opt_logname,mrQ.SunGrid,mrQ.proclus);
     else
         mrQ.SPGR_PDfit_done=1;
     end
-    
+
     save(mrQ.name,'mrQ');
 else
     fprintf('\n load the claculted PD from the coils  M0               \n');
-    
+
 end
 
 
@@ -390,23 +390,23 @@ end
 
 if (mrQ.SPGR_PDBuild_done==0 && mrQ.SPGR_PDfit_done==1)
     fprintf('build the WF map   form PD fit            ');
-    
-    
+
+
     mrQ.opt=mrQ_buildPD(mrQ.opt_logname);
-    
+
     %[mrQ.PDcombineInfo]=mrQ_BuildCoilsfitsPD(mrQ.spgr_initDir,mrQ.proclass);
     mrQ.SPGR_PDBuild_done=1;
     save(mrQ.name,'mrQ');
 else
     fprintf('load the WF map               ');
-    
+
 end
 
 % calculate VIP TV and SIR
 
 if (mrQ.SPGR_PDBuild_done==1)
     fprintf('\n calculate VIP TV SIR form T1 and WF maps               \n');
-    
+
     [mrQ.AnalysisInfo] = mrQ_VIP(mrQ.spgr_initDir);
     save(mrQ.name,'mrQ');
 end
@@ -415,15 +415,15 @@ end
 fprintf('\n orgenized the maps in a maps directory                \n');
 mapDir=fullfile(mrQ.spgr_initDir,'maps');
 
-            % if we redo it and maps dir is already exsist we will saved
-            % the old before we make a new one
+% if we redo it and maps dir is already exsist we will saved
+% the old before we make a new one
 if (exist(mapDir,'dir'))
     ex=0; num=0;
     while ex==0
         num=num+1;
         mapDirOld=fullfile(mrQ.spgr_initDir,['mapsOld_' num2str(num)] );
         if (~exist(mapDirOld,'dir'))
-            
+
             eval(['! mv ' mapDir  ' ' mapDirOld]);
             ex=1;
         end
@@ -432,8 +432,19 @@ end
 
 mkdir(mapDir);
 
-cmd =(['! mv ' mrQ.spgr_initDir '/T1_map_lsq.nii.gz ' mapDir '/.']) ;
-eval(cmd);
+% If T1_map_lsq.nii.gz exists in the spgr_intDir then move it to the map
+% dir. Otherwise assume that it ended up in mapDirOld and retain this t1
+% map. This would occur if the pd fits were rerun but the t1 fits were not
+if exist(fullfile(mrQ.spgr_initDir, 'T1_map_lsq.nii.gz'),'file')
+    cmd =(['! mv ' mrQ.spgr_initDir '/T1_map_lsq.nii.gz ' mapDir '/.']);
+    eval(cmd);
+else
+    % If a new t1 map was not generated than copy it from the old
+    % directory. We copy rather than move to leave the old directory intact
+    cmd =(['! cp ' mapDirOld '/T1_map_lsq.nii.gz ' mapDir '/.']);
+    eval(cmd);
+end
+
 
 cmd =(['! mv ' mrQ.spgr_initDir '/WF_map.nii.gz ' mapDir '/.']) ;
 eval(cmd);
@@ -447,8 +458,6 @@ eval(cmd);
 cmd =(['! mv ' mrQ.spgr_initDir '/SIR_map.nii.gz ' mapDir '/.']) ;
 eval(cmd);
 
-cmd =(['! mv ' mrQ.spgr_initDir '/cT1SIR_map.nii.gz ' mapDir '/.']) ;
-eval(cmd);
 
 
 mrQ.mapsDir=mapDir;
@@ -456,7 +465,6 @@ mrQ.maps.T1path=fullfile(mapDir,'T1_map_lsq.nii.gz');
 mrQ.maps.WFpath=fullfile(mapDir,'WF_map.nii.gz');
 mrQ.maps.TVpath=fullfile(mapDir,'TV_map.nii.gz');
 mrQ.maps.SIRpath=fullfile(mapDir,'SIR_map.nii.gz');
-mrQ.maps.cT1SIRpath=fullfile(mapDir,'cT1SIR_map.nii.gz');
 
 %done
 mrQ.AnalysisDone=1;
